+++ back/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Routes for load testing with basic auth
 - Number of users field in course table
+- Number of users that have an environment running by course
 
 ## 2.32.2
 ### Fixed
@@ -20,6 +21,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 2.32.0
 ### Changed
 - Determine whether to put courseId-<courseId> or shared-pool label on containers based on if there is a node in the swarm that has the courseId label
+
+### Added 
+- Paths for which permissions will be set before container start
 
 ### Fixed
 - Handle deployments with no sessions attached